Adds a new environment variable : ETSI_014_TEST_SUITE_BASE_CLIENT_URL. Changed existing environment variable : `ETSI_014_TEST_SUITE_BASE_SERVER_URL` instead of `ETSI_014_TEST_SUITE_BASE_URL`
This makes it possible to test the ETSI in case the client and server are running on two separate instances, with a different url/port.

The default value provided in the Makefile have only been duplicated to avoid breaking existing code using this Makefile.

The README.md as well has been updated to reflect the changes.

The .gitignore has been update to avoid pushing IDE configuration.